backend/llm/requirement.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
groq
langchain-groq
beautifulsoup4
fastapi
uvicorn